% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pagespeed.R
\name{get_pagespeed}
\alias{get_pagespeed}
\title{Query the Pagespeed API for a given page}
\usage{
get_pagespeed(ps_url = NULL, key = NULL, filter_tpr = NULL,
  locale = NULL, rule = NULL, screenshot = NULL, snapshots = NULL,
  strategy = NULL, utm_campaign = NULL, utm_source = NULL,
  fields = NULL)
}
\arguments{
\item{ps_url}{The URL which you'd like to analyse}

\item{key}{The API key (not required) which you'll be using to query the Google API}

\item{filter_tpr}{Indicates if third party resources should be filtered out before PageSpeed analysis. (boolean)}

\item{locale}{The locale used to localize formatted results (string)}

\item{rule}{A PageSpeed rule to run; if none are given, all rules are run (string)}

\item{screenshot}{Indicates if binary data containing a screenshot should be included (boolean)}

\item{snapshots}{Indicates if binary data containing snapshot images should be included (boolean)}

\item{strategy}{The analysis strategy (desktop or mobile) to use, and desktop is the default (string)}

\item{utm_campaign}{Campaign name for analytics. (string)}

\item{utm_source}{Campaign source for analytics. (string)}

\item{fields}{Selector specifying which fields to include in a partial response.}
}
\value{
Returns a data frame containing key pagespeed insights for the given request
}
\examples{
jaguar <- get_pagespeed(ps_url = "http://www.jaguar.co.uk", key = "my_key", strategy = "desktop")

See also metrics reference at https://developers.google.com/web/tools/chrome-user-experience-report/#metrics
}
